заменил int на double...

+++ laba 1/Func_image.cpp
@@ -11,13 +11,13 @@ void Func_image::plot_image(std::vector<std::vector<double>> image) {
     plt::show();
 }
 
-void Func_image::plot_image(std::vector<std::vector<float>> image, std::string name, int index) {
+void Func_image::plot_image(std::vector<std::vector<double>> image, std::string name, int index) {
     if (image.size() != 2) {
         std::cout << "Error: image must be 2D" << std::endl;
         return;
     }
-    std::vector<float> x_axis = image[0];
-    std::vector<float> y_axis = image[1];
+    std::vector<double> x_axis = image[0];
+    std::vector<double> y_axis = image[1];
     plt::figure(index);
     plt::plot(x_axis, y_axis);
     plt::title(name);
